+++ pyproject.toml
@@ -17,6 +17,7 @@ python-dotenv = "^0.20.0"
 [tool.poetry.dev-dependencies]
 black = "^22.3.0"
 flake8 = "^4.0.1"
+pytest = "^7.1.2"
 
 [tool.black]
 line-length = 100
